+++ lib/core/data/event/repository/event_repository_impl.dart
@@ -1,11 +1,13 @@
 import 'package:app/core/data/event/dtos/event_cohost_request_dto/event_cohost_request_dto.dart';
 import 'package:app/core/data/event/dtos/event_dtos.dart';
+import 'package:app/core/data/event/dtos/event_join_request_dto/event_join_request_dto.dart';
 import 'package:app/core/data/event/dtos/event_rsvp_dto/event_rsvp_dto.dart';
 import 'package:app/core/data/event/gql/event_mutation.dart';
 import 'package:app/core/data/event/gql/event_query.dart';
 import 'package:app/core/domain/event/entities/event.dart';
 import 'package:app/core/domain/event/entities/event_checkin.dart';
 import 'package:app/core/domain/event/entities/event_cohost_request.dart';
+import 'package:app/core/domain/event/entities/event_join_request.dart';
 import 'package:app/core/domain/event/entities/event_rsvp.dart';
 import 'package:app/core/domain/event/event_repository.dart';
 import 'package:app/core/domain/event/input/accept_event_input/accept_event_input.dart';
@@ -24,6 +26,9 @@ import 'package:app/injection/register_module.dart';
 import 'package:dartz/dartz.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
 import 'package:injectable/injectable.dart';
+import 'package:app/graphql/backend/event/query/get_event_join_request.graphql.dart';
+import 'package:app/graphql/backend/event/mutation/approve_user_join_requests.graphql.dart';
+import 'package:app/graphql/backend/event/mutation/decline_user_join_requests.graphql.dart';
 
 @LazySingleton(as: EventRepository)
 class EventRepositoryImpl implements EventRepository {
@@ -277,5 +282,66 @@ class EventRepositoryImpl implements EventRepository {
             .toList(),
       ),
     );
+  }
+
+  @override
+  Future<Either<Failure, List<EventJoinRequest>>> getEventJoinRequests({
+    required Variables$Query$GetEventJoinRequests input,
+  }) async {
+    final result = await client.query$GetEventJoinRequests(
+      Options$Query$GetEventJoinRequests(
+        variables: input,
+        fetchPolicy: FetchPolicy.networkOnly,
+      ),
+    );
+
+    if (result.hasException) {
+      return Left(Failure.withGqlException(result.exception));
+    }
+    return Right(
+      (result.parsedData?.getEventJoinRequests ?? [])
+          .map(
+            (item) => EventJoinRequest.fromDto(
+              EventJoinRequestDto.fromJson(
+                item.toJson(),
+              ),
+            ),
+          )
+          .toList(),
+    );
+  }
+
+  @override
+  Future<Either<Failure, bool>> approveUserJoinRequest({
+    required Input$ApproveUserJoinRequestsInput input,
+  }) async {
+    final result = await client.mutate$ApproveUserJoinRequests(
+      Options$Mutation$ApproveUserJoinRequests(
+        variables: Variables$Mutation$ApproveUserJoinRequests(
+          input: input,
+        ),
+      ),
+    );
+    if (result.hasException) {
+      return Left(Failure.withGqlException(result.exception));
+    }
+    return Right(result.parsedData?.approveUserJoinRequests ?? false);
+  }
+
+  @override
+  Future<Either<Failure, bool>> declineUserJoinRequest({
+    required Input$DeclineUserJoinRequestsInput input,
+  }) async {
+    final result = await client.mutate$DeclineUserJoinRequests(
+      Options$Mutation$DeclineUserJoinRequests(
+        variables: Variables$Mutation$DeclineUserJoinRequests(
+          input: input,
+        ),
+      ),
+    );
+    if (result.hasException) {
+      return Left(Failure.withGqlException(result.exception));
+    }
+    return Right(result.parsedData?.declineUserJoinRequests ?? false);
   }
 }